requirements.txt
insightface
fastapi
opencv-python
onnxruntime
numpy
pymongo
asyncio
uvicorn
python-dotenv